Refactor functions into vue composables
